# Backend/jiufang-project/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 九方智能查询系统 API
  description: |
    九方智能查询系统是一个基于AI Agent的自然语言查询系统，支持用户通过自然语言查询ERP数据，
    并提供定时报告、预警规则、数据导出等功能。
    
    ## 认证方式
    - 使用JWT Bearer Token认证
    - 登录接口获取JWT令牌
    - 令牌有效期24小时
    
    ## 权限控制
    - admin: 系统管理员，拥有所有权限
    - manager: 管理层，拥有查询和导出权限
    - executive: 执行层，拥有查询权限
    
    ## 响应格式
    所有接口统一使用以下响应格式：
    ```json
    {
      "code": 200,
      "message": "success",
      "data": {}
    }
    ```
  version: 1.0.0
  contact:
    name: 九方技术团队
    email: tech@jiufang.com
servers:
  - url: https://api.jiufang.com/v1
    description: 生产环境
  - url: https://api-dev.jiufang.com/v1
    description: 开发环境
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
...
paths:
...
  /favorites:
    post:
      tags:
        - Favorite
      summary: 创建收藏
      description: 创建查询记录收藏
      operationId: createFavorite
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - record_id
              properties:
                record_id:
                  type: string
                  pattern: '^\d{19}$'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags:
        - Favorite
      summary: 查询收藏列表
      description: 分页查询用户的收藏列表
      operationId: getFavoriteList
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /favorites/{favorite_id}:
    delete:
      tags:
        - Favorite
      summary: 删除收藏
      description: 删除收藏
      operationId: deleteFavorite
      security:
        - bearerAuth: []
      parameters:
        - name: favorite_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== 对话收藏模块 ====================
...
  /export:
    post:
      tags:
        - Export
      summary: 导出查询结果
      description: 导出查询结果为Excel或CSV格式
      operationId: exportQueryResult
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - record_id
                - format
              properties:
                record_id:
                  type: string
                  pattern: '^\d{19}$'
                format:
                  type: string
                  enum: [xlsx, csv]
      responses:
        '200':
          description: 导出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /export/records:
    get:
      tags:
        - Export
      summary: 查询导出记录
      description: 查询用户的导出记录列表
      operationId: getExportRecords
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportRecordListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== 定时报告模块 ====================
...
  /feedbacks:
    post:
      tags:
        - Feedback
      summary: 创建反馈
      description: 创建用户反馈
      operationId: createFeedback
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 1000
                type:
                  type: string
                  enum: [bug, suggestion, question]
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags:
        - Feedback
      summary: 查询反馈列表
      description: 分页查询反馈列表（仅系统管理员可操作）
      operationId: listFeedbacks
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /feedbacks/{id}:
    get:
      tags:
        - Feedback
      summary: 查询反馈详情
      description: 查询反馈详情（仅系统管理员可操作）
      operationId: getFeedback
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== 操作日志模块 ====================
...
components:
  schemas:
    SuccessResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: 错误码
          example: 40001
        message:
          type: string
          description: 错误信息
          example: "参数错误"
...
    FavoriteResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - id
            - record_id
            - created_at
          properties:
            id:
              type: string
              pattern: '^\d{19}$'
            record_id:
              type: string
              pattern: '^\d{19}$'
            created_at:
              type: string
              format: date-time
    FavoriteListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - total
            - page
            - page_size
            - favorites
          properties:
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
            favorites:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    pattern: '^\d{19}$'
                  record_id:
                    type: string
                    pattern: '^\d{19}$'
                  query_text:
                    type: string
                  created_at:
                    type: string
                    format: date-time
...
    ExportResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - id
            - file_url
            - created_at
          properties:
            id:
              type: string
              pattern: '^\d{19}$'
            file_url:
              type: string
              format: uri
            created_at:
              type: string
              format: date-time
    ExportRecordListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - total
            - page
            - page_size
            - records
          properties:
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
            records:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    pattern: '^\d{19}$'
                  file_url:
                    type: string
                    format: uri
                  format:
                    type: string
                    enum: [xlsx, csv]
                  created_at:
                    type: string
                    format: date-time
...
    FeedbackResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - id
            - content
            - created_at
          properties:
            id:
              type: string
              pattern: '^\d{19}$'
            content:
              type: string
            type:
              type: string
              enum: [bug, suggestion, question]
            user_id:
              type: string
              pattern: '^\d{19}$'
            created_at:
              type: string
              format: date-time
    FeedbackListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - total
            - page
            - page_size
            - feedbacks
          properties:
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
            feedbacks:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    pattern: '^\d{19}$'
                  content:
                    type: string
                  type:
                    type: string
                    enum: [bug, suggestion, question]
                  user_id:
                    type: string
                    pattern: '^\d{19}$'
                  created_at:
                    type: string
                    format: date-time
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40001
            message: "参数错误"
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40101
            message: "未授权访问"
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40301
            message: "权限不足"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40401
            message: "资源不存在"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证